+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 ## DART 6
 
-### [DART 6.9.1 (2019-XX-XX)](https://github.com/dartsim/dart/milestone/59?closed=1)
+### [DART 6.9.2 (2019-07-XX)](https://github.com/dartsim/dart/milestone/60?closed=1)
+
+* Dynamics
+
+  * Allowed constraint force mixing > 1: [#1371](https://github.com/dartsim/dart/pull/1371)
+
+### [DART 6.9.1 (2019-06-06)](https://github.com/dartsim/dart/milestone/59?closed=1)
+
+* Collision
+
+  * Added default constructor to RayHit: [#1345](https://github.com/dartsim/dart/pull/1345)
 
 * dartpy